+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # CHANGELOG
 
 ## 0.9.5 (unreleased)
-- CHANGE: `encrypt-file` changed to `encrypt-path` (to match the `encrypt_path` parameter of the `encrypt` method)
+- CHANGED:
+    * `encrypt-file` changed to `encrypt-path` (to match the `encrypt_path` parameter of the `encrypt` method)
+    * parameter swap from `.encrypt(sign=None, key=True)` to `.encrypt(key=True, sign=None)` due to S/MIME that does not take the key from the GPG keyring but needs the certificate to be specified every time. 
 - S\MIME signing (`M2Crypto` instead of `smime` package)
+- unit tests + travis build status
 - SMTP
     * if SMTP/INI file is given as a relative path and not found at CWD, we try program directory (useful when importing envelope as a library from another program)
     * TLS security supported